migrate tests to use result instead of unwrap
motivation: in this way tests are like in-library code and there is no
differenciation between test code and library code in terms of style,
where errors are returned with `?` instead of `unwrap`. Same for
examples. This was already done partially like this in some of the
ProtoGalaxy tests, now on almost all the folding-schemes tests.